refactor VVC encoding
