+++ libheif/codecs/vvc.cc
@@ -24,6 +24,8 @@
 #include <cassert>
 #include <iomanip>
 #include <utility>
+#include <libheif/api_structs.h>
+
 
 Error Box_vvcC::parse(BitstreamRange& range)
 {
@@ -546,4 +548,70 @@ Error parse_sps_for_vvcC_configuration(const uint8_t* sps, size_t size,
   config->constant_frame_rate = 1; // is constant (TODO: where do we get this from)
 
   return Error::Ok;
+}
+
+
+Result<ImageItem::CodedImageData> ImageItem_VVC::encode(const std::shared_ptr<HeifPixelImage>& image,
+                                                        struct heif_encoder* encoder,
+                                                        const struct heif_encoding_options& options,
+                                                        enum heif_image_input_class input_class)
+{
+  CodedImageData codedImage;
+
+  auto vvcC = std::make_shared<Box_vvcC>();
+  codedImage.properties.push_back(vvcC);
+
+
+  heif_image c_api_image;
+  c_api_image.image = image;
+
+  struct heif_error err = encoder->plugin->encode_image(encoder->encoder, &c_api_image, input_class);
+  if (err.code) {
+    return Error(err.code,
+                 err.subcode,
+                 err.message);
+  }
+
+  int encoded_width = 0;
+  int encoded_height = 0;
+
+  for (;;) {
+    uint8_t* data;
+    int size;
+
+    encoder->plugin->get_compressed_data(encoder->encoder, &data, &size, NULL);
+
+    if (data == NULL) {
+      break;
+    }
+
+
+    const uint8_t NAL_SPS = 15;
+
+    uint8_t nal_type = 0;
+    if (size>=2) {
+      nal_type = (data[1] >> 3) & 0x1F;
+    }
+
+    if (nal_type == NAL_SPS) {
+      Box_vvcC::configuration config;
+
+      parse_sps_for_vvcC_configuration(data, size, &config, &encoded_width, &encoded_height);
+
+      vvcC->set_configuration(config);
+    }
+
+    switch (nal_type) {
+      case 14: // VPS
+      case 15: // SPS
+      case 16: // PPS
+        vvcC->append_nal_data(data, size);
+        break;
+
+      default:
+        codedImage.append_with_4bytes_size(data, size);
+    }
+  }
+
+  return codedImage;
 }
